Work at Home Thursday
used a calculation for FF based on how it is used by the sparkmax library, because voltage is supposed to correspond to a specific RPM when there is no load on the motor.
removed arbitrary feed forward controller in favor of this change as well. kept values and placeholders in constants just in case